+++ ngboost/__init__.py
@@ -7,8 +7,15 @@ except ImportError:
     from importlib_metadata import version
 
 from .api import NGBClassifier, NGBRegressor, NGBSurvival
+from .helpers import load_ngboost_model
 from .ngboost import NGBoost
 
-__all__ = ["NGBClassifier", "NGBRegressor", "NGBSurvival", "NGBoost"]
+__all__ = [
+    "NGBClassifier",
+    "NGBRegressor",
+    "NGBSurvival",
+    "NGBoost",
+    "load_ngboost_model",
+]
 
 __version__ = version(__name__)